+++ gradle.properties
@@ -2,14 +2,14 @@
 org.gradle.jvmargs=-Xmx3G
 org.gradle.parallel=true
 #Minecraft
-minecraft_version=1.20.5-rc2
+minecraft_version=1.21
 #Mod
 archives_base_name=special-model-loader
 mod_display_name=SpecialModelLoader
 mod_version=1.2.0
 #Dependencies
-loader_version=0.15.10
-fabric_version=0.97.3+1.20.5
+loader_version=0.15.11
+fabric_version=0.100.0+1.21
 #Maven
 maven_group=dev.felnull
 maven_put_url=https://maven.felnull.dev/
