+++ src/main/java/dev/felnull/specialmodelloader/api/model/SpecialBaseLoader.java
@@ -2,7 +2,7 @@ package dev.felnull.specialmodelloader.api.model;
 
 import com.google.gson.JsonObject;
 import dev.felnull.specialmodelloader.impl.SpecialModelLoader;
-import net.fabricmc.fabric.api.client.model.ModelProviderException;
+//import net.fabricmc.fabric.api.client.model.ModelProviderException;
 import net.minecraft.client.resources.model.UnbakedModel;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.resources.ResourceManager;
@@ -10,7 +10,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 public interface SpecialBaseLoader {
-    @Nullable UnbakedModel loadModel(@NotNull ResourceManager resourceManager, @NotNull JsonObject modelJson) throws ModelProviderException;
+    @Nullable UnbakedModel loadModel(@NotNull ResourceManager resourceManager, @NotNull JsonObject modelJson) throws IllegalStateException;
 
     @Nullable String getId();
