+++ src/main/java/dev/felnull/specialmodelloader/impl/model/ForgeObjModelCompat.java
@@ -6,7 +6,7 @@ import net.minecraft.resources.ResourceLocation;
 import org.apache.commons.lang3.tuple.Pair;
 
 public final class ForgeObjModelCompat {
-    private static final ResourceLocation FORGE_OBJ = new ResourceLocation("forge", "obj");
+    private static final ResourceLocation FORGE_OBJ = ResourceLocation.fromNamespaceAndPath("forge", "obj");
 
     public static Pair<ResourceLocation, ObjModelOption> getObjModelData(JsonObject modelJson) {
         if (modelJson == null)
@@ -15,7 +15,7 @@ public final class ForgeObjModelCompat {
         if (!modelJson.has("loader") || !modelJson.get("loader").isJsonPrimitive() || !modelJson.getAsJsonPrimitive("loader").isString() || !FORGE_OBJ.toString().equals(modelJson.get("loader").getAsString()))
             return null;
 
-        var model = new ResourceLocation(modelJson.get("model").getAsString());
+        var model = ResourceLocation.parse(modelJson.get("model").getAsString());
         return Pair.of(model, ObjModelOption.parse(modelJson));
     }
 }
